# lib/api-zod/src/generated/types/asset.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * NETHERWORLD TERMINAL - Crypto & Forex Trading Signal Bot API
 * OpenAPI spec version: 0.1.0
 */
import type { AssetType } from './assetType';

export interface Asset {
  id: number;
  symbol: string;
  type: AssetType;
  isActive: boolean;
}