# lib/api-zod/src/generated/types/profile.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * NETHERWORLD TERMINAL - Crypto & Forex Trading Signal Bot API
 * OpenAPI spec version: 0.1.0
 */

export interface Profile {
  id: number;
  creatorName: string;
  /** @nullable */
  bio: string | null;
  /** @nullable */
  contactInfo: string | null;
  /** @nullable */
  socialLinks: string | null;
  /** @nullable */
  photoUrl: string | null;
}