# lib/api-zod/src/generated/types/sessionInfo.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * NETHERWORLD TERMINAL - Crypto & Forex Trading Signal Bot API
 * OpenAPI spec version: 0.1.0
 */

export interface SessionInfo {
  authenticated: boolean;
  role: string;
  username: string;
}